Make IKOS available in the spaceros image (#99).
IKOS is one of the tools that Space ROS includes to perform static analysis of
ROS 2 applications. Although ament_ikos is provided with the docker image, IKOS
itself is not, rendering the former useless and leaving it to users to compile
and install IKOS prior to being able to use it.

This commit modifies the spaceros Earthfile to install IKOS globally in the
system.